fix to post all required interpreter method entry and re-entry notifications
the native stub was failing to notify an entry

the rethrow handler was failing to notify a re-entry

the call stub was notifying a re-entry in all cases whereas it only
needs to do so when a normal return occurs since returns to the call
stub via exception rethrows already do a notify.

n.b. the latter fix requires inserting a notify instruction directly
after the branch to the method entry thus shifting the saved return
entry address by one instruction. a normal return executes the notify
whereas a rethrow starts at the saved address i.e. the instruction
following the notify. this does not require any change to the rethrow
handler code -- which simply jumps to the saved address. it /does/
affect the stack traversal code -- which detects a top-level (call
stub) frame by comparing the return address in the frame with the
saved address. When notification is enabled the stack walk must
subtract 4 (size of notify instruction) from the saved address before
doing the compare.