make partialSubtypeCheck pass expected registers to check_klass_subtype_slow_path
the implementation of partialSubtypeCheck uses MacroAssembler method
check_klass_subtype_slow_path. since the latter spills any of the
registers it does not expect before use we force the allocator to use
the ones it does expect. if this involves shuffles or saves then we
will be no worse off and we might be able to use them free anyway.